+++ src/OpenRiaServices.DomainServices.Client.Web/Framework/Web/WebDomainClientFactory.cs
@@ -19,6 +19,7 @@ namespace OpenRiaServices.DomainServices.Client.Web
         /// Initializes a new instance of the <see cref="WebDomainClientFactory" /> class.
         /// </summary>
         public WebDomainClientFactory()
+            : base("binary")
         {
         }
 
@@ -51,17 +52,6 @@ namespace OpenRiaServices.DomainServices.Client.Web
             }
 
             return factory;
-        }
-
-        /// <summary>
-        /// Get an <see cref="EndpointAddress" /> that identifies the server endpoint
-        /// </summary>
-        /// <param name="endpoint">Absolute service URI without protocol suffix such as "/binary"</param>
-        /// <param name="requiresSecureEndpoint"><c>true</c> if communication must be secured, otherwise <c>false</c></param>
-        /// <returns><see cref="EndpointAddress"/> where target uri has the protocol suffix ("/binary") set </returns>
-        protected override EndpointAddress CreateEndpointAddress(Uri endpoint, bool requiresSecureEndpoint)
-        {
-            return new EndpointAddress(new Uri(endpoint.OriginalString + "/binary", UriKind.Absolute));
         }
 
         /// <summary>